Project/requirements.txt
modal>=0.64.0
torch>=2.2.0
transformers>=4.40.0
tokenizers>=0.19.0
sentencepiece>=0.2.0
protobuf>=3.20.0
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorboard>=2.15.0
pyyaml>=6.0
tqdm>=4.66.0
wandb>=0.16.0
emoji>=2.10.0
pyvi>=0.1.1
